story fixes
