+++ components/side-drawer/src/vwc-side-drawer-base.ts
@@ -117,7 +117,7 @@ export class VWCSideDrawerBase extends LitElement {
 			</aside>
 
 			<div class="side-drawer--app-content">
-			<slot name="app-content"></slot>
+				<slot name="app-content"></slot>
 			</div>
 
 			${scrim}
