+++ components/side-drawer/src/vwc-side-drawer.scss
@@ -44,6 +44,7 @@ $inline-size: var(#{variables.$side-drawer-inline-size}, 280px);
 	&.side-drawer--dismissible.side-drawer--open,
 	&:not(.side-drawer--dismissible):not(.side-drawer--modal) {
 		+ .side-drawer--app-content {
+			@include typography.typography-cat-shorthand(body-2);
 			margin-inline-start: #{$inline-size};
 			::slotted(vwc-top-app-bar),
 			::slotted(vwc-top-app-bar-fixed) {
@@ -55,7 +56,7 @@ $inline-size: var(#{variables.$side-drawer-inline-size}, 280px);
 
 .side-drawer--top-bar {
 	$gutter: var(#{variables.$side-drawer-padding-top-bar}, #{$padding-default});
-	@include typography.typography-cat-shorthand(subtitle-2);
+	@include typography.typography-cat-shorthand(subtitle-1);
 	display: flex;
 	flex-shrink: 0;
 	align-items: center;
